requirements.txt
python-dotenv>=1.0.1
fastapi>=0.115.12
google-genai>=1.20.0
google-adk>=1.3.0
PyPDF2>=3.0.1
python-docx>=1.1.2
uvicorn[standard]>=0.32.1
python-multipart>=0.0.6
litellm>=1.0.0
openai>=1.0.0
pymupdf>=1.24.0
